scripts/py/azure_function/requirements.txt
PyMuPDF
msal
requests
dotenv
azure-functions
